Clean stale comments + consolidate test files
Pre-merge tidy on #26.  No behavior change, no API change.

1. stylesheet.css `.lightbox-video-container`:
   - Drop the misleading "reserving ~180px for caption + close button"
     comment.  The real sizing now comes from fitVideoToLightbox() with
     M = max(24px, 5vmin); the static CSS is just a pre-JS-flash
     fallback.
   - Drop the "No margin: auto" comment.  It explains an absence — the
     thing isn't there, the explanation belongs in commit history, not
     in code.

2. Test file consolidation:
   - Move all of tests/test_video_lightbox_aspect_ratios.py into
     tests/test_video_lightbox.py with VIEWPORTS expanded from
     {1280×720, 1920×1080} to also include {320×568, 768×1024}.
   - Delete tests/test_video_lightbox_aspect_ratios.py.
   - Net matrix: 4 sources × 4 viewports × 5 invariants = 80 tests
     (was 16 + 40 = 56 across two files with overlapping 16:9 cells).
   - Removes "PR #27" references that won't resolve cleanly in main's
     history post-squash; the invariants are now described directly.

3. tests/test_video_lightbox_captions.py:
   - Drop the "Fixed by PR #27" reference for the same reason.  Test
     describes the contract it pins, not the historical fix-PR.

4. LAYOUT_REVIEW.md:
   - Add a status preface marking the items resolved since the review
     was written (P0.1 video lightbox crash; P1 margin: auto; Open Q
     video centering).  Everything else still applies — TODOS.md is
     the live state.
